CP #3

+++ app/controllers/calculations_controller.rb
@@ -67,7 +67,7 @@ class CalculationsController < ApplicationController
     
     def process_square
         
-        @user_number = params["a_number"].to_i
+        @user_number = params["a_number"].to_f
         @squared_number = @user_number**2
         
         render("calculations/square_results_template.html.erb")
@@ -83,7 +83,7 @@ class CalculationsController < ApplicationController
     
     def process_square_root
         
-        @user_number = params["a_number"].to_i
+        @user_number = params["a_number"].to_f
         @sqrt_number = Math.sqrt(@user_number)
         
         render("calculations/square_root_results_template.html.erb")
